Fix #820 (Unhandled exception "TypeError: 'NoneType' object is not callable")
* Fix bug: Unhandled exception "TypeError: 'NoneType' object is not callable" in tools.py function __log_keyring_warning (#820).
   Logging thread removed and logger module correctly initialized as fix. Is "Heisenbug" so 100 % retesting was not possible.
* Refactor: Solved circular dependency between tools.py and logger.py to fix #820
