+++ CHANGES
@@ -3,10 +3,13 @@ Back In Time
 Version 1.4.2-dev (development of upcoming release)
 * Fix bug: RTE: module 'qttools' has no attribute 'initate_translator' with encFS when prompting the user for a password (#1553).
 * Fix bug: Schedule dropdown menu used "minutes" instead of "hours".
+* Fix bug: Unhandled exception "TypeError: 'NoneType' object is not callable" in tools.py function __log_keyring_warning (#820).
+           Logging thread removed and logger module correctly initialized as fix. Is "Heisenbug" so 100 % retesting was not possible.
 * Build: Use PyLint in unit testing to catch E1101 (no-member) errors.
 * Build: Activate PyLint warning W1401 (anomalous-backslash-in-string).
 * Build: Add codespell config.
 * Translation: Minor modifications in source strings and updating language files.
+* Refactor: Solved circular dependency between tools.py and logger.py to fix #820
 
 Version 1.4.1 (2023-10-01)
 * Dependency: Add "qt translations" to GUI runtime dependencies (#1538).
